sample-run.txt: made dirdup call more precise
Currently, if the user had already initialized poda with other
hampers already index, those would be considered on the cat, and
would make the output differ from the sample run. Furthermore,
poda-dirdupes.py would take long to run, making it appear that
it hanged.

With this change, at least this will only happen if the other
hampers are called samplehamper, which should be rare enough
to eliminate other possibilities.